requirements.txt
streamlit
openai
galileo
python-dotenv
langchain
langchain-core
langchain-openai
langgraph
langchain-pinecone
pinecone-client
langchain-text-splitters
langchain-community
